Atualizações.

+++ app/Http/Controllers/ProjectController.php
@@ -5,6 +5,7 @@ namespace CodeEducation\Http\Controllers;
 use CodeEducation\Repositories\ProjectRepository;
 use CodeEducation\Services\ProjectService;
 use Illuminate\Http\Request;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectController extends Controller
 {
@@ -20,7 +21,7 @@ class ProjectController extends Controller
 
     public function index()
     {
-        return $this->repository->with(['client', 'user'])->all();
+        return $this->repository->with(['client', 'user'])->findWhere(['owner_id' => Authorizer::getResourceOwnerId()]);
     }
 
     public function store(Request $request)
@@ -30,16 +31,33 @@ class ProjectController extends Controller
 
     public function show($id)
     {
+        if(!$this->checkProjectOwner($id)) {
+            return ['error' => 'Acesso negado!'];
+        }
+
         return $this->repository->find($id);
     }
 
     public function update(Request $request, $id)
     {
+        if(!$this->checkProjectOwner($id)) {
+            return ['error' => 'Acesso negado!'];
+        }
+
         return $this->service->update($request->all(), $id);
     }
 
     public function destroy($id)
     {
+        if(!$this->checkProjectOwner($id)) {
+            return ['error' => 'Acesso negado!'];
+        }
+
         return $this->service->delete($id);
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        return $this->repository->isOwner($projectId, Authorizer::getResourceOwnerId());
     }
 }
